+++ src/utils/constants.ts
@@ -6,43 +6,33 @@ import { ZanixWorkerConnector } from 'connectors/worker.ts'
 import { ZanixAsyncmqConnector } from 'connectors/asyncmq.ts'
 
 /**
- * Reserved port for default SOCKET server
+ * Default port for SOCKET server
  */
 export const SOCKET_PORT = 20201
 /**
- * Reserved port for default STATIC server
+ * Default port for STATIC server
  */
 export const STATIC_PORT = 20202
 /**
- * Reserved port for default GQL server
+ * Default port for GQL server
  */
 export const GRAPHQL_PORT = 20203
 /**
- * Reserved port for default admin REST server
+ * Default port for admin REST server
  */
 export const ADMIN_REST_PORT = 30248
 /**
- * Reserved port for default admin GQL server
+ * Default port for admin GQL server
  */
 export const ADMIN_GRAPHQL_PORT = 30249
 /**
- * Reserved port for default admin SOCKET server
+ * Default port for admin SOCKET server
  */
 export const ADMIN_SOCKET_PORT = 30250
 /**
- * Reserved port for default admin STATIC server
+ * Default port for admin STATIC server
  */
 export const ADMIN_STATIC_PORT = 30251
-
-export const RESERVED_PORTS = [
-  STATIC_PORT,
-  SOCKET_PORT,
-  GRAPHQL_PORT,
-  ADMIN_GRAPHQL_PORT,
-  ADMIN_REST_PORT,
-  ADMIN_SOCKET_PORT,
-  ADMIN_STATIC_PORT,
-]
 
 /**
  * Content header for http JSON application